Remove Monitor header from Accounts
