+++ src/components/screens/monitor/accounts/index.js
@@ -1,14 +1,56 @@
-/* istanbul ignore file */
 import React from 'react';
 import { compose } from 'redux';
 import { withTranslation } from 'react-i18next';
-import { useSelector } from 'react-redux';
+import Box from '../../../toolbox/box';
+import BoxHeader from '../../../toolbox/box/header';
+import BoxContent from '../../../toolbox/box/content';
+import withResizeValues from '../../../../utils/withResizeValues';
+import Table from '../../../toolbox/table';
+import styles from './accounts.css';
+import header from './tableHeader';
+import AccountRow from './accountRow';
 import withData from '../../../../utils/withData';
-import MonitorAccounts from './accounts';
 import liskServiceApi from '../../../../utils/api/lsk/liskService';
-import NotAvailable from '../notAvailable';
 
-const ComposedAccounts = compose(
+const Accounts = ({
+  accounts,
+  networkStatus,
+  title,
+  t,
+}) => {
+  /* istanbul ignore next */
+  const handleLoadMore = () => {
+    accounts.loadData({ offset: accounts.data.length });
+  };
+  const supply = networkStatus.data.supply;
+  const canLoadMore = accounts.meta ? accounts.meta.count === 30 : false;
+
+  return (
+    <Box main isLoading={accounts.isLoading} className="accounts-box">
+      <BoxHeader>
+        <h1>{title}</h1>
+      </BoxHeader>
+      <BoxContent className={styles.content}>
+        <Table
+          data={accounts.data}
+          isLoading={accounts.isLoading}
+          row={AccountRow}
+          loadData={handleLoadMore}
+          header={header(t)}
+          additionalRowProps={{ supply }}
+          error={accounts.error}
+          canLoadMore={canLoadMore}
+        />
+      </BoxContent>
+    </Box>
+  );
+};
+
+Accounts.defaultProps = {
+  title: '',
+};
+
+export default compose(
   withData(
     {
       accounts: {
@@ -29,17 +71,6 @@ const ComposedAccounts = compose(
       },
     },
   ),
+  withResizeValues,
   withTranslation(),
-)(MonitorAccounts);
-
-const AccountsMonitor = () => {
-  const network = useSelector(state => state.network);
-
-  return (
-    liskServiceApi.getLiskServiceUrl(network) === null
-      ? <NotAvailable />
-      : <ComposedAccounts />
-  );
-};
-
-export default AccountsMonitor;
+)(Accounts);
